don't use remote machine's hostname in wmic
This fixes a bug in wmic where all addresses would be resolved against
the lmhosts file, wins, /etc/hosts file, bcast, etc. However, there are
some situations where wmic is used against a remote machine that is not
on the local network.

Instead of using a hostname that will probably never resolve, and in
lieu of terrible LD_PRELOAD hacks, just override whatever the bindings
claim the hostname is and use the original ip address passed in through
the command line.

Also, credit goes to Kartic Krishnamurthy for figuring out that this was
a bug in wmic/winexe/samba and not something wrong with Windows
firewalls.